feat(commands): mine, tweak,refactor: bot
