+++ bot/bot.ts
@@ -1,11 +1,9 @@
 import tmi, { Client, ChatUserstate } from "tmi.js";
-import yaml from "js-yaml";
-import fs from "node:fs";
 import { CommandBase } from "./commandBase";
 
 export const PATH =
   "C:\\Files\\Servers\\PaperServerDev1\\plugins\\TrySomethingDevAmazingPlugin\\config.yml";
-const ADMIN_NAME = "TrySomethingDev";
+export const ADMIN_NAME = "TrySomethingDev";
 
 // Define configuration options
 const opts = {
@@ -16,14 +14,14 @@ const opts = {
   channels: ["trysomethingdev"],
 };
 
+export const channelName = opts.channels[0];
+
 // Create a client with our options
 const client = new tmi.client(opts);
+client.connect();
 export type ClientType = Client;
 
 export const commandBase = new CommandBase(client);
-
-// Connect to Twitch:
-client.connect();
 
 export interface MessageHandler {
   (channel: string, context: ChatUserstate, msg: string, self: boolean): void;
@@ -44,53 +42,16 @@ const onMessageHandler: MessageHandler = (
   const commandExists = commandBase.findCommand(commandName);
   if (!commandExists) return;
 
-  commandBase.runCommand(commandName, [msg]);
+  commandBase.runCommand(commandName, client, context, msg.split(" "));
 
-  if (
-    commandName.toUpperCase() === "!MINECLEAR" &&
-    context["display-name"] === ADMIN_NAME
-  ) {
-  }
-
-  // If the command is known, let's execute it
-  if (commandName === "!mine") {
-    //Add name of sender to ""
-
-    let doc = yaml.load(fs.readFileSync(PATH, "utf8"));
-    if (doc.ChattersThatWantToPlay.includes(context["display-name"])) {
-      client.say(
-        target,
-        context["display-name"] + `, your name is already on the list`,
-      );
-    } else {
-      doc.ChattersThatWantToPlay.push(context["display-name"]);
-      doc.General.Greeting = msg;
-      fs.writeFile(PATH, yaml.dump(doc), (err) => {
-        if (err) {
-          console.log(err);
-        }
-        client.say(
-          target,
-          context["display-name"] + `, your name has been added to the list.`,
-        );
-      });
-    }
-
-    console.log(`* Executed ${commandName} command`);
-  }
+  console.log(`* Executed ${commandName} command`);
 };
 
 // Register our event handlers (defined below)
 client.on("message", onMessageHandler);
 client.on("connected", onConnectedHandler);
 
-// Function called when the "dice" command is issued
-function rollDice() {
-  const sides = 6;
-  return Math.floor(Math.random() * sides) + 1;
-}
-
 // Called every time the bot connects to Twitch chat
-function onConnectedHandler(addr, port) {
+function onConnectedHandler(addr: string, port: number) {
   console.log(`* Connected to ${addr}:${port}`);
 }
